+++ CHANGELOG.markdown
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+Development
+-----------
+
+* Fixed preloading failure in IEs 6 and 7 (GH-243).
+* Non-looping multi-rows fixed (GH-244).
+* Negative speed now rolling in correct direction (GH-241).
+* Compatibility with Visual Studio Windows 8 apps added (GH-263).
+* Cache teardown removal fixed (GH-265).
+* Improved performance of debugging monitor.
+
+
 Version 1.3
 -----------